Fix missed flaky test case of REINDEX TABLE/INDEX
	commit 2ab75d6f7d5cc5eaf8ff1db9de07bbfb69aecb3e
	Author: Adam Lee <adam.lee@enterprisedb.com>
	Date:   Thu Aug 14 15:15:30 2025 +0800

	    Fix flaky tests of REINDEX TABLE/INDEX

The previous commit missed one case, which is fixed here. I double-checked and
there should be no more.

The root cause is the same as before: after a legitimate functional change,
REINDEX TABLE on partitioned tables can no longer be run inside an explicit
transaction block, which prevented the tests from working as intended and
caused them to be flaky.