[ORCA] Fix a segmentation fault when append statistics to a group.
When appending statistics to a group, it is first checked whether statistics
already exist in the group. If it exists, then new statistics are appended to
the existing one using the AppendStats method. If there are no statistics in the
group, the existence of statistics in the duplicate group is also checked. When
appending statistics to an existing one, we take the existing statistics of the
group (or its duplicate), create a copy of it, and add statistics to this copy,
and release the old one. If the group does not have its own statistics and the
duplicate statistics are used. Then we would add statistics to the duplicate
group and try to release the statistics of the current group, which is NULL,
which leads to a segmentation fault.

Fix this by calling the AppendStats method on the duplicate.